feat: implement GET endpoint for fetching all chats and update POST response status; refactor chat retrieval in chatId route
