+++ app/api/chats/[chatId]/route.ts
@@ -1,15 +1,18 @@
 import { NextResponse } from 'next/server';
 import { db } from '@/db';
 import { chats } from '@/db/schema';
-import { desc } from 'drizzle-orm';
+import { eq } from 'drizzle-orm';
 
-export async function GET() {
+export async function GET(_req: Request, ctx: RouteContext<'/api/chats/[chatId]'>) {
   try {
-    const chatsData = await db.select().from(chats).orderBy(desc(chats.updatedAt));
+    const { chatId } = await ctx.params;
+    const [chat] = await db.select().from(chats).where(eq(chats.id, chatId)).limit(1);
 
-    console.log('chats data: ', chatsData);
+    if (!chat) {
+      return NextResponse.json({ error: 'Chat not found' }, { status: 404 });
+    }
 
-    return NextResponse.json(chatsData, { status: 200 });
+    return NextResponse.json(chat, { status: 200 });
   } catch (error) {
     console.log('GET /api/chats/[chatId] failed:', error);
